requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.1
django-resized==0.3.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
djoser==0.5.4
drfdocs==0.0.11
gunicorn==19.6.0
mysqlclient==1.3.9
olefile==0.44
packaging==16.8
Pillow==4.0.0
PyJWT==1.4.2
pyparsing==2.1.10
six==1.10.0
